+++ frameworks/GalileoControl.framework/Headers/GalileoControl.h
@@ -1,0 +1,18 @@
+//  Created by Chris Harding on 1/17/13.
+//  Copyright (c) 2013 MOTRR LLC. All rights reserved.
+//
+/** This header is a convenience header for importing all the Galileo public interface classes.
+ */
+
+#import "GalileoCommon.h"
+//
+#import "GalileoDelegate.h"
+#import "Galileo.h"
+//
+#import "VelocityControlDelegate.h"
+#import "VelocityControl.h"
+//
+#import "PositionControlDelegate.h"
+#import "PositionControl.h"
+//
+#import "FirmwareManager.h"